Add setup.py installer
